chore: [PR6] split types to clear SwiftLint length errors
No behavior change — move code into extensions / a top-level
OSPushSubscriptionImpl so file_length and type_body_length stay under
error thresholds.
